services/gateway/requirements.txt
# =============================================================================
# TRJM Gateway - Python Dependencies
# =============================================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
starlette==0.36.3

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-ldap==3.4.4
cryptography==42.0.2

# HTTP Client (for LLM providers)
httpx==0.26.0
openai==1.12.0

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# File Processing
python-docx==1.1.0
pypdf2==3.0.1
reportlab==4.1.0
extract-msg==0.48.4
python-magic==0.4.27
chardet==5.2.0

# Rate Limiting
slowapi==0.1.9

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==8.2.0
pytest-asyncio==0.24.0
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
ruff==0.2.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
orjson==3.9.13
tenacity==8.2.3
